pasdfo

+++ app/Http/Controllers/DocumentosController.php
@@ -15,8 +15,13 @@ class DocumentosController extends Controller
 
     public function indexAdminTransp()
     {
-        $documentos = Documentos::search()->orderBy('nombre_documento')->paginate(300);
-        return view('admintransparencia', compact('documentos'));
+        if(Auth::guard('admin')->check()){
+            $documentos = Documentos::search()->orderBy('nombre_documento')->paginate(300);
+            return view('admintransparencia', compact('documentos'));
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
 
     public function create()
@@ -31,25 +36,30 @@ class DocumentosController extends Controller
 
     public function storeActa(Request $request)
     {
+        if(Auth::guard('admin')->check()){
 
-        $this->validate($request, [
-            'nombre_acta'=> 'required',
-            'tipo_documento' => 'required',
-            'fecha_acta'=> 'required',
-            'url_acta'=> 'required|url|active_url',
+            $this->validate($request, [
+                'nombre_acta'=> 'required',
+                'tipo_documento' => 'required',
+                'fecha_acta'=> 'required',
+                'url_acta'=> 'required|url|active_url',
 
-        ]);
+            ]);
 
-        $documentos = new Documentos;
-        $documentos->nombre_documento = $request->input('nombre_acta');
-        $documentos->url_documento = $request->input('url_acta');
-        $documentos->fecha_documento = $request->input('fecha_acta');
-        $documentos->tipo_documento = $request->input('tipo_documento');
+            $documentos = new Documentos;
+            $documentos->nombre_documento = $request->input('nombre_acta');
+            $documentos->url_documento = $request->input('url_acta');
+            $documentos->fecha_documento = $request->input('fecha_acta');
+            $documentos->tipo_documento = $request->input('tipo_documento');
 
-        $documentos->save();
-    
-    
-        return redirect('/admin/transparencia')->with('success','Acta añadida correctamente.');
+            $documentos->save();
+        
+        
+            return redirect('/admin/transparencia')->with('success','Acta añadida correctamente.');
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
 
     public function show($id)
@@ -64,8 +74,13 @@ class DocumentosController extends Controller
 
     public function editActa($id)
     {
-        $documento = Documentos::find($id);
-        return view('editaractas')->with('documento', $documento);
+        if(Auth::guard('admin')->check()){
+            $documento = Documentos::find($id);
+            return view('editaractas')->with('documento', $documento);
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
 
     public function update(Request $request, $id)
@@ -75,32 +90,42 @@ class DocumentosController extends Controller
 
     public function updateActa(Request $request, $id)
     {
-        $this->validate($request, [
-            'nombre_acta'=> 'required',
-            'tipo_documento' => 'required',
-            'fecha_acta'=> 'required',
-            'url_acta'=> 'required|url|active_url',
+        if(Auth::guard('admin')->check()){
+            $this->validate($request, [
+                'nombre_acta'=> 'required',
+                'tipo_documento' => 'required',
+                'fecha_acta'=> 'required',
+                'url_acta'=> 'required|url|active_url',
 
-        ]);
+            ]);
 
-        $documentos = Documentos::find($id);
-        $documentos->nombre_documento = $request->input('nombre_acta');
-        $documentos->url_documento = $request->input('url_acta');
-        $documentos->fecha_documento = $request->input('fecha_acta');
-        $documentos->tipo_documento = $request->input('tipo_documento');
+            $documentos = Documentos::find($id);
+            $documentos->nombre_documento = $request->input('nombre_acta');
+            $documentos->url_documento = $request->input('url_acta');
+            $documentos->fecha_documento = $request->input('fecha_acta');
+            $documentos->tipo_documento = $request->input('tipo_documento');
 
-        $documentos->save();
-    
-    
-        return redirect('/admin/transparencia')->with('success','Acta actualizada correctamente.');
+            $documentos->save();
+        
+        
+            return redirect('/admin/transparencia')->with('success','Acta actualizada correctamente.');
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
 
     public function destroy($id)
     {
-        $documento = Documentos::find($id);
-        $documento->delete();
-        
-        return redirect('/admin/transparencia')->with('success', 'El acta se ha borrado con éxito.');
+        if(Auth::guard('admin')->check()){
+            $documento = Documentos::find($id);
+            $documento->delete();
+            
+            return redirect('/admin/transparencia')->with('success', 'El acta se ha borrado con éxito.');
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
     
 }
