+++ app/Http/Controllers/ElemQrController.php
@@ -15,9 +15,13 @@ class ElemQrController extends Controller
 
     public function indexAdmin()
     {
-        
-        $elementos = ElemQr::search()->orderBy('nombre_elemqr')->paginate(20);
-        return view('admineleqr', compact('elementos'));
+        if(Auth::guard('admin')->check()){    
+            $elementos = ElemQr::search()->orderBy('nombre_elemqr')->paginate(20);
+            return view('admineleqr', compact('elementos'));
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
     
     public function create()
@@ -27,27 +31,32 @@ class ElemQrController extends Controller
 
     public function store(Request $request)
     {
-        //return $request->input('name');
-        //obliga a que los campos de name y email se llenen,
-        //si no no hace submit.
-        $this->validate($request, [
-            'nombre_eleqr'=> 'required',
-            'descripcion_eleqr'=> 'required',
-            'url_imagen_eleqr'=> 'required|url|active_url',
+        if(Auth::guard('admin')->check()){
+            //return $request->input('name');
+            //obliga a que los campos de name y email se llenen,
+            //si no no hace submit.
+            $this->validate($request, [
+                'nombre_eleqr'=> 'required',
+                'descripcion_eleqr'=> 'required',
+                'url_imagen_eleqr'=> 'required|url|active_url',
 
-        ]);
+            ]);
 
-        //  return 123;
-        //create new ReservaVisitaGuiada
-        $elemqr = new ElemQr;
-        $elemqr->nombre_elemqr = $request->input('nombre_eleqr');
-        $elemqr->descripcion_elemqr = $request->input('descripcion_eleqr');
-        $elemqr->url_img_elemqr = $request->input('url_imagen_eleqr');
-    
-        //save message
-        $elemqr->save();
-    
-        return redirect('/admin/elemqr')->with('success','Elemento añadido con éxito.');
+            //  return 123;
+            //create new ReservaVisitaGuiada
+            $elemqr = new ElemQr;
+            $elemqr->nombre_elemqr = $request->input('nombre_eleqr');
+            $elemqr->descripcion_elemqr = $request->input('descripcion_eleqr');
+            $elemqr->url_img_elemqr = $request->input('url_imagen_eleqr');
+        
+            //save message
+            $elemqr->save();
+        
+            return redirect('/admin/elemqr')->with('success','Elemento añadido con éxito.');
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
 
    
@@ -70,34 +79,49 @@ class ElemQrController extends Controller
      */
     public function edit($id)
     {
-        $elemento = ElemQr::find($id);
-        return view('editarelemqr')->with('elemento', $elemento); 
+        if(Auth::guard('admin')->check()){
+            $elemento = ElemQr::find($id);
+            return view('editarelemqr')->with('elemento', $elemento);
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        } 
     }
 
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
-            'nombre_eleqr'=> 'required',
-            'descripcion_eleqr'=> 'required',
-            'url_imagen_eleqr'=> 'required|url|active_url',
+        if(Auth::guard('admin')->check()){
+            $this->validate($request, [
+                'nombre_eleqr'=> 'required',
+                'descripcion_eleqr'=> 'required',
+                'url_imagen_eleqr'=> 'required|url|active_url',
 
-        ]);
+            ]);
 
-        $elemqr = ElemQr::find($id);
-        $elemqr->nombre_elemqr = $request->input('nombre_eleqr');
-        $elemqr->descripcion_elemqr = $request->input('descripcion_eleqr');
-        $elemqr->url_img_elemqr = $request->input('url_imagen_eleqr');
- 
-        $elemqr->save();
+            $elemqr = ElemQr::find($id);
+            $elemqr->nombre_elemqr = $request->input('nombre_eleqr');
+            $elemqr->descripcion_elemqr = $request->input('descripcion_eleqr');
+            $elemqr->url_img_elemqr = $request->input('url_imagen_eleqr');
     
-        return redirect('/admin/elemqr')->with('success','Elemento modificado con éxito.');
+            $elemqr->save();
+        
+            return redirect('/admin/elemqr')->with('success','Elemento modificado con éxito.');
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
 
     public function destroy($id)
     {
-        $elemento = ElemQr::find($id);
-        $elemento->delete();
-        
-        return redirect('/admin/elemqr')->with('success', 'El Elemento QR se ha eliminado correctamente.');
+        if(Auth::guard('admin')->check()){
+            $elemento = ElemQr::find($id);
+            $elemento->delete();
+            
+            return redirect('/admin/elemqr')->with('success', 'El Elemento QR se ha eliminado correctamente.');
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
 }
